# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_server.APIResponse:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  internal_server.CreateTokenRequest:
    properties:
      description:
        type: string
      expires_in:
        description: seconds
        type: integer
      name:
        type: string
      scopes:
        items:
          type: string
        type: array
    type: object
  internal_server.TokenResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      revoked:
        type: boolean
      scopes:
        items:
          type: string
        type: array
      token:
        description: Only included on creation
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Multi-terminal AI orchestration platform for task scheduling, agent
    management, and intelligent routing.
  title: AI Orchestration Platform API
  version: "2.1"
paths:
  /auth/tokens:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_server.TokenResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
      security:
      - BearerAuth: []
      summary: List user API tokens
      tags:
      - auth
    post:
      consumes:
      - application/json
      parameters:
      - description: Token configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_server.CreateTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_server.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new API token
      tags:
      - auth
  /auth/tokens/{id}:
    delete:
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete an API token
      tags:
      - auth
  /auth/tokens/{id}/revoke:
    post:
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
      security:
      - BearerAuth: []
      summary: Revoke an API token
      tags:
      - auth
  /system/backups:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
      security:
      - BearerAuth: []
      summary: List all backups
      tags:
      - backup
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a manual backup
      tags:
      - backup
  /system/backups/{filename}/restore:
    post:
      parameters:
      - description: Backup filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
      security:
      - BearerAuth: []
      summary: Restore from a backup
      tags:
      - backup
  /system/backups/{filename}/verify:
    post:
      parameters:
      - description: Backup filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
      security:
      - BearerAuth: []
      summary: Verify a backup checksum
      tags:
      - backup
  /system/backups/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.APIResponse'
      security:
      - BearerAuth: []
      summary: Get backup statistics
      tags:
      - backup
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"